+++ src/main/java/faang/school/postservice/service/comment/CommentService.java
@@ -3,6 +3,7 @@ package faang.school.postservice.service.comment;
 import faang.school.postservice.dto.comment.CommentDto;
 import faang.school.postservice.dto.comment.CommentEvent;
 import faang.school.postservice.exception.DataValidationException;
+import faang.school.postservice.helper.UserCacheWriter;
 import faang.school.postservice.mapper.comment.CommentMapper;
 import faang.school.postservice.model.Comment;
 import faang.school.postservice.model.Post;
@@ -12,11 +13,15 @@ import faang.school.postservice.repository.CommentRepository;
 import faang.school.postservice.service.post.PostService;
 import faang.school.postservice.validator.comment.CommentValidator;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.Comparator;
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutorService;
 
 @Service
 @AllArgsConstructor
@@ -27,6 +32,11 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final CommentEventPublisher commentEventPublisher;
     private final KafkaCommentPostEventPublisher kafkaCommentPostEventPublisher;
+    private final UserCacheWriter userCacheWriter;
+
+    @Autowired
+    @Qualifier("writeToCacheThreadPool")
+    private ExecutorService writeToCacheThreadPool;
 
     public Comment findEntityById(long id) {
         return commentRepository.findById(id)
@@ -46,6 +56,8 @@ public class CommentService {
 
         publishCommentCreationEvent(comment);
         sendCommentPostEvent(comment);
+        long authorId = comment.getAuthorId();
+        CompletableFuture.runAsync(() -> userCacheWriter.cacheUser(authorId), writeToCacheThreadPool);
 
         return commentMapper.toDto(comment);
     }
